Converted to double-based location system

+++ GradeTrackerPro/src/gradetrackerpro/graphics/AGraphicsEntity.java
@@ -2,22 +2,22 @@ package gradetrackerpro.graphics;
 
 
 public abstract class AGraphicsEntity implements IRenderable{
-  private int x;
-  private int y;
+  private double x;
+  private double y;
   private int width;
   private int height;
   private boolean visibility;
-  public AGraphicsEntity(int x, int y, int width, int height){
+  public AGraphicsEntity(double x, double y, int width, int height){
     this.x=x;
     this.y=y;
     this.width=width;
     this.height=height;
     this.visibility=true;
   }
-  public int getX(){
+  public double getX(){
     return this.x;
   }
-  public int getY(){
+  public double getY(){
     return this.y;
   }
   public int getWidth(){
@@ -26,7 +26,7 @@ public abstract class AGraphicsEntity implements IRenderable{
   public int getHeight(){
     return this.height;
   }
-  public void setLocation(int x, int y){
+  public void setLocation(double x, double y){
     this.x=x;
     this.y=y;
   }
