+++ GradeTrackerPro/src/gradetrackerpro/graphics/screens/ScreenChooseCourseName.java
@@ -47,8 +47,8 @@ public class ScreenChooseCourseName extends JPanel implements IReceiver, ITrigge
 	}
 	public void ping(String title, String[] data){
 		if(title.equals("mouse-data")){
-			int x = Integer.parseInt(data[0]);
-			int y = Integer.parseInt(data[1]);
+			int x = (int)Double.parseDouble(data[0]);
+			int y = (int)Double.parseDouble(data[1]);
 			int event = Integer.parseInt(data[2]);
 			this.nameTextBox.ping("mouse-data", data);
 			this.goButton.mouseAction(x, y, event);
